+++ packages/core/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-- Support retype from contract or interface to address. ([#687](https://github.com/OpenZeppelin/openzeppelin-upgrades/issues/687))
+- Support retype from contract, interface, struct or enum to address. ([#687](https://github.com/OpenZeppelin/openzeppelin-upgrades/issues/687))
 
 ## 1.20.4 (2022-11-03)